requirements.txt
flask==2.3.3
msal==1.22.0
requests==2.31.0
google-generativeai==0.3.1
langchain==0.0.267
langchain-google-genai==0.0.5
python-dotenv==1.0.0
chromadb==0.4.15
beautifulsoup4==4.12.2